Not allowing unknown properties in the API objects.
Rrather than just stripping them; this may be better for validation and ensuring only expected objects make it into the api/db. But on the flip side, stripping may be fine, since that is more forward-compatible and allows more fields to be added.

Also removing unneeded log line (since we are logging any validation errors anyways).