Adding a 31-day expiration for the storage of raw responses in S3.
The objects will be automatically removed from the bucket by AWS S3 after the expiration period.

Also setting a prefix for responses, so that the rule will only apply to objects stored under that prefix (this will allow us to have the choice of easily supporting data that doesn't expire in the future).